Add a script for generating the NDK documentation.
This is not going to exactly match the view on devsite since that goes
through a different layer of tooling and is themed differently, but
it's a decent approximation for checking most formatting issues.

Longer term, we need to get prebuilts for doxygen into the tree so we
can revive `m ndk-docs`.

Test: ./build_ndk_docs.py
Bug: http://b/115644359
Change-Id: Ib1533a8f45731a1ae2fbec79d463d17e9124beaa