Implement global state for files with Vuex
With increasing interaction between components it makes sense to have a global state so any component can update and read the global state instead of passing updates up and down the Vue component tree. In particular this is useful so that the current timeline time can be updated from anywhere easily with the changes reflected in all components.

Test: N/A
Change-Id: Ie099973b8a36915c7e93e4d716b0f92e82ab28b4