sdk.atree += libbinder_ndk C++ helpers
libbinder_ndk includes a set of C++ headers which are for all practical
purposes coupled with the AIDL compiler. Since NDK and SDK releases do
not line up, and some developers are known to use even old versions of
the SDK with the newest NDK, subtle C++ changes to this library may
break compatibility with the aidl compiler.

Both to reduce the possibility we accidentally break developers as well
as increase the types of changes that can be made, the plan is to always
ship libbinder_ndk C++ headers alongside AIDL.

In the future, this will come with a removal of these headers from the
NDK (with error messages pointing developers to use the version that is
built and tested with their version of aidl).

Bug: 160624671
Test: 'm sdk' and verify contents in out/host/*/sdk/sdk/**
Change-Id: I6dcdfce5314863663548f0a4523df6e96448ee23