Resolve hidden connectivity dependency
ConnectivityManager and NetwokUtils are parts of incoming
Connectivity mainline module. The hidden methods or members are
no longer accessible outside the module. Caller should either
remove or replace the usage.

The override hidden callbacks and INET_CONDITION_ACTION intent
are not really used in the code. Remove the unncessary usage.

The network id is a part of Network.toString(). The network id
is also not necessary.

Bug: 182859030
Test: m
Change-Id: I3a75afb86f5b60ea530b50771bd76887ccf9f271