Allow cargo2android.py to use a config file instead of command-line options.
The command-line options to c2a are already long and will only get
longer as new arguments are added.  This allows it to load them from a
config file.  For simplicity of implementation, the config file
supplements the command-line arguments and uses the same format they
do, although it ideally will replace them over time.

This also adds an option to dump the current set of command-line
arguments to a config file and exit.  This is intended to ease
migration over to the config file: for a given crate, we can call c2a
with its command-line arguments plus this new flag to generate the
config file and then use just that instead.

Test: Dump and load some config files.
Change-Id: I63b29dd20bcff5d5832dbd380d7c6eb273547ed0