Unify DiffGenerator with flicker objects
Previously the DiffGenerator on flicker used an independent structure.
This was necessary because iwnscope used flicker objects for WM and its
own objects for SF traces.

Now that both traces use flicker objects this is no longer necessary, so
remove this code to simplify the codebase and improve performance

Also, remove unused code from transform_sf because the objects are now
entirely parsed thorugh flicker and no operations are necessary in
winscope

Finally, fix eslint issues on the files (camel case, lack of semicolon,
unused variables) to improve readability

Bug: 162300507
Bug: 162302603
Test: build winscope and open a few WM and SF traces
Change-Id: Idb2c816767f0e6941d7e1242ca2ca6b2799dd4e0