Move the '--partial' flag and output configuration to the backend.
Previously, the front end will generate a string for '--partial' flag.
Now the front end will only record the partial partitions that are going
to be updated as an array, and the backend will convert the array into a
string for CLI tool.

The output now is also calculated in the backend,

Test: mannual tested.
Change-Id: Ib5604e52b0c0d0cdd2bf85c71479b3133479da74