Bounds representation improvements.
Highlighted rects change background color to a red tint to make them
easier to see. The new displays property on an entry is now shown as a
blue-dashed rect on the SF trace - should still be able to click on
other rects to highlight them. Old SF traces (without display property)
should not be affected. Boundless displays now constrained to 1.3 times
display size.

Bug: b/198306246
Test: Upload trace to x20 (see bug).
Change-Id: Id80e539cd5f302754146d0bb1708784e3a57067c