[1] Introduce beta feature flag for new Ime panels
Add a new toolbar at the top of the front page of Winscope app.
This toolbar will show a list of beta features and checkboxes for
users to choose whether to use the beta feature.

For now, the only beta feature is the new / improved UI version
of the ImeTrace panels (to be implemented in the next few CLs).

Bug: 236226833
Test: manual (on local winscope build)

Change-Id: If64aff9e899363188010633758e36be474666984