Add tslint
Add tslint with google3 config (google3/javascript/typescript/tslint).

Skipped google3 custom rules that need to be built and added to the project.

If needed, with some extra effort it should be possible
to get the google3 rules to build/run also in the Winscope environment.
However, it is probably simpler and quicker to run (and maybe fix?)
those checks on critique once we port Winscope to google3.

Fix: b/264505907
Test: npm run tslint:check / npm run tslint:fix
Change-Id: If3a79de673503e0c06541f0329163dee93849c3c