cargo_embargo: improve the empty test detection check
The changes I made in https://r.android.com/2804478 aren't correct. The
string extracted via the regex isn't the package name. The first part
seems to be the crate name (which for integration tests is not the same
as the package name and might be duplicated across packages within a
workspace, so it isn't a good namespace by itself) and the hash comes
from the "-C extra-filename" rustc arg.

Bug: 242733327
Test: ran on crosvm
Change-Id: I26977054b23c12e7785a438d5ea0ff1e92780b2f