Fullscreen support in VDM Client
Replacing the host's immersive mode. Volume down goes back
to the main activity. Supports rotation.

Adding pause/resume for all displays so we don't send frames
when the client activity is not in foreground.

Fixing the key event dispatch, so the top focused display has
first dibs but if not tracking focus, the event reaches the client
app.

Fixing the IFTTT paths because they don't seem to work like in g3.

Fix: 314729842
Test: manual
Change-Id: I153c595e688b79339b3b9db9936e4cc33ccef487