Suggest running gdb-remote again on failure
The lldbclient script has a race condition where it starts lldb-server
on the device and runs the client in quick succession, resulting in the
client attempting to perform the handshake before the server has had
time to fully start.

Until a mechanism is found to retry the connection automatically or
otherwise block on lldb-server being ready, add a comment suggesting the
user retry gdb-remote manually. This occurs most frequently on the
RISC-V Cuttlefish emulator, where the first attempt of `gdb-remote`
tends to fail and the second attempt almost always passes.

Test: python3 development/scripts/lldbclient.py --port 1234 -r /system/bin/date
Test: Verified print of comment when running client script.
Change-Id: I2c80344a586ad1b54ab9759b40e4e9db46f49084